Julien Pommier's fixes for msvc compiler compatibility and msvc project files

+++ src/getfem_fem_level_set.cc
@@ -104,7 +104,7 @@ namespace getfem {
 	mls_of_convex(c.convex_num());
       scalar_type v = eval(c.xref());
       // z[i] = (v > 1e-9) ? '+' : ((v < -1e-9) ? '-' : '0');
-      z[i] = (v > 0.) ? '+' : '-';
+      z[(size_t)i] = (v > 0.) ? '+' : '-';
     }
     unsigned cnt = 0;
     for (unsigned d = 0; d < dofzones.size(); ++d) {
